redaction

+++ trunk/ii02810/task_02/src/model.cpp
@@ -1,4 +1,5 @@
 #include "model.h"
+#include <cmath>
 #include <iostream>
 
 std::vector<float> Linemodel(float y, float u, int steps, float a, float b) {
@@ -18,7 +19,7 @@ std::vector<float> UnLinemodel(const Params& p, float y, float y0) {
     }
 
     for (int i = 0; i < p.steps; ++i) {
-        auto y_next = (p.a * y - p.b * y0 * y0 + p.c * p.u + p.d * std::sinf(p.u));
+        auto y_next = (p.a * y - p.b * y0 * y0 + p.c * p.u + p.d * std::sin(p.u));
         result.push_back(y_next);
         y0 = y;
         y = y_next;
